requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==2015.04.28
gevent==1.0.2
greenlet==0.4.7
pika==0.9.14
redis==2.10.3
requests==2.7.0
six==1.9.0
tornado==4.2
websocket-client==0.25.0
ws4py==0.3.4
wsgiref==0.1.2
